xskxmak

+++ lib/ewma.cc
@@ -25,6 +25,6 @@ DirectEWMAX<stability_shift, scale>::update_zero_period(unsigned period)
   else {
     int compensation = 1 << (stability_shift - 1); // round off
     for (; period > 0; period--)
-      _avg += (-_avg + compensation) >> stability_shift;
+      _avg += static_cast<int>(-_avg + compensation) >> stability_shift;
   }
 }
